removed silences from final textgrid
