+++ output-SRT-file.srt
@@ -1,0 +1,17 @@
+1
+00:00:00,000 --> 00:07:08,569
+[Silence]
+
+2
+00:07:08,569 --> 00:07:11,209
+I am an igloo
+- and I'm on a new line
+
+3
+00:07:11,209 --> 00:07:11,309
+[Silence]
+
+4
+00:07:11,309 --> 00:07:16,660
+And the SRT is working
+
